axios

+++ api-proyecto/controllers/productos.controller.js
@@ -1,22 +1,24 @@
-const productoModels = require('../models/producto.models');
+//const productoModels = require('../models/producto.models');
+const fs = require('fs');
+const path = require('path');
 const Producto = require('../models/producto.models');
-const imagen=require("../utils/image")
+const imagen = require('../utils/image');
 
 async function createProducto(req, res) {
     const producto = new Producto(req.body);
 
-    if(req.files.imagep){
-        const imagePath=imagen.getfilePath(req.files.imagep)
-        producto.imagep=imagePath
-    
+    if (req.files.imagep) {
+        const imagePath = imagen.getfilePath(req.files.imagep)
+        producto.imagep = imagePath
+
     }
 
     try {
         const datos = await producto.save();
         res.status(200).send(datos)
     } catch (error) {
-       console.log(error);
-       //res.status(500).send({ message: "error al guardar los producto" })
+        console.log(error);
+        //res.status(500).send({ message: "error al guardar los producto" })
 
     }
 }
@@ -30,15 +32,26 @@ async function getProducto(req, res) {
     }
 }
 async function delProducto(req, res) {
-    const { id } = req.params
+    const { id } = req.params;
     try {
-        await Producto.findByIdAndDelete(id)
-        res.status(200).send({ message: 'Dato eliminado correctamente' })
+        const producto = await Producto.findById(id);
+        if (!producto) {
+            return res.status(404).send({ message: "Producto no encontrado" });
+        }
+
+        const imagePath = path.join(__dirname, '..', producto.imagep);
+
+        fs.unlink(imagePath, (error) => {
+            if (error) {
+                return res.status(500).send({ message: "Error al eliminar la imagen" });
+            }
+        });
+
+        await Producto.findByIdAndDelete(id);
+        res.status(200).send({ message: "Producto eliminado" });
     } catch (error) {
-        res.status(500).send({ message: 'No se ha podido eliminar el dato' })
+        res.status(500).send({ message: "Error al eliminar" });
     }
-
-
 }
 async function updateProducto(req, res) {
     const { id } = req.params
